requirements.txt
fastapi
uvicorn
python-multipart
langchain
google-generativeai
